requirements.txt
BeautifulSoup4
Django==3.2.7
django-filter==2.4.0
django-heroku==0.3.1
django-polymorphic==3.0.0
django-picklefield==3.0.1
django-rest-framework==0.1.0
django-rest-polymorphic==0.1.9
django-widget-tweaks==1.4.8
django-webpack-loader==1.3.0
django-loginas==0.3.10
nltk==3.5
gensim==3.8.2
smart-open==1.10.0
pandas==1.0.4
numpy==1.19
scipy==1.5
spacy==2.2.4
spacy-udpipe==0.3.2
celery==5.0.5
django-celery-results==2.0.1
celery-progress==0.1.0
statsmodels==0.12.2
psycopg2-binary==2.9.1
redis==3.5.3
django-redis==5.0.0
python-memcached==1.59
pymemcache==3.5.0
kombu==5.1.0
podium-nlp==0.1.1
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.1+cu113
transformers==4.13.0
rank-bm25==0.2.2
pyro-ppl==1.8.0